feat: display movable seats on chart
