+++ server/templates/student.html.j2
@@ -17,7 +17,7 @@
       <a href="{{ url_for('room', exam=exam, id=seat.room.id, seat=seat.id) }}">
         {{ seat.room.display_name }} {{ seat.display_name }}
       </a><br><br>
-      <a target="_blank" href="{{ url_for('single_seat', seat_id=seat.id) }}">
+      <a target="_blank" href="{{ url_for('student_single_seat', seat_id=seat.id) }}">
         Sharable Link to This Seat
       </a>
     {% else %}
