+++ tests/fixtures/seat.json
@@ -53,6 +53,12 @@
         "room_id": 1,
         "fixed": false,
         "attributes": ["Righty"]
+      },
+      {
+        "id": 7,
+        "room_id": 1,
+        "fixed": false,
+        "attributes": ["Lefty"]
       }
     ]
   }
